Revert "Add new @ shortcut for HEAD"
This reverts commit b25e1125ee03c83cccb6d51a489ac3f3df0c94cc, as it
does not just apply to "@" (and forms with modifiers like @{u}
applied to it), but also affects e.g. "refs/heads/@/foo", which it
shouldn't.

The basic idea of giving a short-hand might be good, and the topic
can be retried later, but let's revert to avoid affecting existing
use cases for now for the upcoming release.